136958 make a module for folding drawio
